+++ src/commands/hardis/org/retrieve/sources/metadata.ts
@@ -54,7 +54,7 @@
       this.ux.log(powerkitRes.message);
 
       // Retrieve metadatas
-      await MetadataUtils.retrieveMetadatas(packageXml, folder, false, [], this, debug);
+      await MetadataUtils.retrieveMetadatas(packageXml, folder, false, [], {}, this, debug);
 
       // Copy to destination
       await fs.copy(path.join(folder, 'unpackaged'), path.resolve(folder));